ordered repo structure and improved the cmake

+++ src/io_executor.cc
@@ -173,7 +173,7 @@ namespace hedge::async
             if(submit < 0)
                 throw std::runtime_error("io_uring_submit: "s + strerror(-submit));
 
-            if(ready != submit) // todo might remove
+            if(ready != static_cast<unsigned int>(submit)) // todo might remove
             {
                 // std::cerr << "Warning: io_uring_submit ready != submit. in flight: "
                 //   << _in_flight_requests.size() << ", ready: " << ready << " submit: " << submit << " requests: " << requests.size() << std::endl;
